# Use Vectors instead of a Dict for nearby offsets