--- a/src/spaces/grid_multi.jl
+++ b/src/spaces/grid_multi.jl
@@ -5,9 +5,9 @@ struct GridSpace{D,P} <: AbstractGridSpace{D,P}
     stored_ids::Array{Vector{Int},D}
     extent::NTuple{D,Int}
     metric::Symbol
-    offsets_at_radius::Dict{Int,Vector{NTuple{D,Int}}}
-    offsets_within_radius::Dict{Int,Vector{NTuple{D,Int}}}
-    offsets_within_radius_no_0::Dict{Int,Vector{NTuple{D,Int}}}
+    offsets_at_radius::Vector{Vector{NTuple{D,Int}}}
+    offsets_within_radius::Vector{Vector{NTuple{D,Int}}}
+    offsets_within_radius_no_0::Vector{Vector{NTuple{D,Int}}}
     indices_within_radius_tuple::Dict{NTuple{D,Int},Vector{NTuple{D,Int}}}
 end
 spacesize(space::GridSpace) = space.extent
@@ -78,9 +78,9 @@ function GridSpace(
         stored_ids,
         d,
         metric,
-        Dict{Int,Vector{NTuple{D,Int}}}(),
-        Dict{Int,Vector{NTuple{D,Int}}}(),
-        Dict{Int,Vector{NTuple{D,Int}}}(),
+        Vector{Vector{NTuple{D,Int}}}(),
+        Vector{Vector{NTuple{D,Int}}}(),
+        Vector{Vector{NTuple{D,Int}}}(),
         Dict{NTuple{D,Int},Vector{NTuple{D,Int}}}(),
     )
 end