This should make the retrieval of the offsets faster.....

The PR removes the dictionaries used to hold the offsets for calculating nearby stuff and substitute them with vectors. The access to a vector should be faster, and so this should help performance

Still needs benchmarks