@Tortar 2023-09-29
No problem, I will not bother you with this, there are more important changes going on, will try to resolve myself, and if I'm totally lost I will ask you for help