@Tortar 2023-09-30
ok, after the simplification allocations seem to have disappeared, will do a brief macro-benchmark to see how much of an improvement it is