@Tortar 2023-09-30
Indeed, the improvement is pretty substantial, running the Game of Life simulation

<details> <summary> Benchmark </summary>

```julia
using Agents, Random, BenchmarkTools

@agent Automaton GridAgent{2} begin end

function build_model(alive_probability, dims, metric = :chebyshev)
    space = GridSpaceSingle(dims; metric, periodic=false)
    status = zeros(Bool, dims)
    new_status = zeros(Bool, dims)
    rules = (2, 3, 3, 3)
    properties = (; rules, status, new_status)
    model = UnremovableABM(Automaton, space; properties, rng = Xoshiro())
    for pos in Agents.positions(model)
        if rand(model.rng) < alive_probability
            status[pos...] = true
        end
    end
    return model
end

function model_step(model)
    new_status = model.new_status
    status = model.status
    @inbounds for pos in Agents.positions(model)
        n = alive_neighbors(pos, model)
        if status[pos...] == true && (n ≤ model.rules[4] && n ≥ model.rules[1])
            new_status[pos...] = true
        elseif status[pos...] == false && (n ≥ model.rules[3] && n ≤ model.rules[4])
            new_status[pos...] = true
        else
            new_status[pos...] = false
        end
    end
    status .= new_status
    return
end

function alive_neighbors(pos, model)
    c = 0
    @inbounds for near_pos in nearby_positions(pos, model)
        if model.status[near_pos...] == true
            c += 1
        end
    end
    return c
end

function run_model_agents()
    model = build_model(0.5, (100, 100))
    Agents.step!(model, dummystep, model_step, 1000)
end

@benchmark run_model_agents()

```
</details>

before:

```julia
julia> @benchmark run_model_agents()
BenchmarkTools.Trial: 22 samples with 1 evaluation.
 Range (min … max):  223.142 ms … 280.279 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     227.121 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   230.362 ms ±  12.257 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▁ ▃█
  ▇█▄██▁▄▄▁▄▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▁
  223 ms           Histogram: frequency by time          280 ms <

 Memory estimate: 1.32 MiB, allocs estimate: 34950.

```
after:

```julia
julia> @benchmark run_model_agents()
BenchmarkTools.Trial: 31 samples with 1 evaluation.
 Range (min … max):  160.820 ms … 172.073 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     162.463 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   163.448 ms ±   2.549 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁  ▄ ▁ ▁█       ▁  ▁
  █▁▆█▆█▆██▁▁▁▆▆▆▆█▆▆█▆▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▆ ▁
  161 ms           Histogram: frequency by time          172 ms <

 Memory estimate: 1.32 MiB, allocs estimate: 34944.

```

it is 30% faster!!